Corrected errors in other int->float changes

+++ src/main/java/com/crowsofwar/avatar/common/command/NodeAbilityGet.java
@@ -78,7 +78,7 @@ public class NodeAbilityGet extends NodeFunctional {
 				player);
 		if (data != null) {
 			
-			int xp = data.getAbilityData(ability).getXp();
+			float xp = data.getAbilityData(ability).getXp();
 			AvatarChatMessages.MSG_ABILITY_GET.send(call.getFrom(), player,
 					ability.getClass().getSimpleName(), xp);
